+++ app/routes/webhooks.jsx
@@ -1,6 +1,7 @@
 import { authenticate } from "../lib/shopify.server";
 import sessionService from "../services/sessionService.server";
 import merchantService from "../services/merchantService.server";
+import settingsService from "../services/settingsService.server";
 
 export const action = async ({ request }) => {
   const { topic, shop, session, admin } = await authenticate.webhook(request);
@@ -16,6 +17,8 @@ export const action = async ({ request }) => {
         try {
           await sessionService.deleteSession(shop);
           await merchantService.deleteMerchant(shop);
+          await settingsService.deleteSettings(shop);
+
           console.log("APP_UNINSTALLED: ", shop)
         } catch (error) {
           console.log("error:", error);